fixed unknown strlen return type making if statement useless

+++ sockets/3. Crowd Control/src/client_handler.c
@@ -109,7 +109,7 @@ void * handle_client_requests(void * client_)
     message[strlen(req)] = '\0';
 
     // Send message to destination client
-    if(write(clients[dest_client_index].conn, message, strlen(message)) < strlen(message)) // write() failed
+    if(write(clients[dest_client_index].conn, message, strlen(message)) < (int) strlen(message)) // write() failed
     {
       perror("write() failed");
       fprintf(stderr, "wrote %d/%d bytes from clients %.3s to %.3s. skipping to next client request.\n");
